+++ internal/server/dto/response/mother_service.go
@@ -9,8 +9,8 @@ type MotherService struct {
 	CreatedAt                time.Time  `json:"created_at"`
 	UpdatedAt                time.Time  `json:"updated_at"`
 	Name                     string     `json:"name"`
-	ExceptionRate            float64    `json:"exception_rate"`
-	ResponseDelayRate        float64    `json:"response_delay_rate"`
+	ExceptionRate            int        `json:"exception_rate"`
+	ResponseDelayRate        int        `json:"response_delay_rate"`
 	ResponseDelayDuration    *int       `json:"response_delay_duration"`
 	RandomResponseDelayMin   *int       `json:"random_response_delay_min"`
 	RandomResponseDelayMax   *int       `json:"random_response_delay_max"`
